committing new branch where scaling difficulty will be implemented. as of the moment the idea is that the user will start with a smaller set of the total characters and as they get more correct it will add more slowly until full profficiency is displayed.
in the previous version all characters were shown to user which is very difficult to learn and so a gradual increase in different characters will help familiarize users to them.

will attempt to makee it so those of the same root are all added together ie, all vowels first and then ka, ki, ku, ke, ko as a second difficulty and so on and so forth.